requirements.txt
matplotlib==3.5.1
numpy==1.20.0
pandas==1.3.4
torch==1.10.0+cu113
torch-explain==0.6.5
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
networkx==2.6.3
shap==0.40.0
pyvis==0.1.9
rdkit-pypi==2021.9.4
karateclub==1.2.3